refactor: bank menu refactoring
