+++ client/.eslintrc.json
@@ -15,7 +15,6 @@
     "import/prefer-default-export": "off",
     "react/jsx-props-no-spreading": "off",
     "react/destructuring-assignment": "off",
-    // "assignmenteslintreact/destructuring-assignment": "off",
     "react/require-default-props": 0,
     "react/forbid-prop-types": 0,
     "no-var": 1,
